+++ nextgisweb/wfsclient/migration/4892a21e-original-datatype.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "4892a21e" } ***/
+
+ALTER TABLE layer_field_wfsclient_layer DROP COLUMN orig_datatype;